fix feature repayment and feature

+++ src/main/java/com/example/capstone/entity/RepaymentPlan.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
 
-import java.sql.Timestamp;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -24,14 +24,14 @@ import java.util.UUID;
 public class RepaymentPlan {
 
     @Id
-    @Column (name = "plan_id")
+    @Column(name = "plan_id")
     private UUID planId;
 
     @Column(name = "user_id")
     private UUID userId;
 
-    @Column(name ="monthly_budget")
-    private double monthlyBudget;
+    @Column(name = "monthly_budget")
+    private BigDecimal monthlyBudget;
 
     @Column(name = "strategy_id")
     private UUID strategyId;
